tweaks to color

+++ framework.h
@@ -131,10 +131,10 @@ typedef struct vfVector
 
 typedef struct vfColor
 {
-	int r;
-	int g;
-	int b;
-	int a;
+	uint8_t r;
+	uint8_t g;
+	uint8_t b;
+	uint8_t a;
 } vfColor;
 
 typedef struct vfTransform
